Potentially Working With Issues
* Added TODOs for important remaining work
* Removed functionality not to be supported in first version
* Temporarily disabled deathlink as it will be almost free to implement

An issue is appearing that prevents testing of the client and server
locally. This seems like an issue on my end and not with any of the
written code.

Values _are_ being read from the game for play time and health, but I
have serious doubts that they're correct. This is the current largest
hurdle, followed by the code that will award items to the player.

A flag is still necessary to determine if the game has been cleared.
This seems like it should be on Gwyn's death, or perhaps when an ending
has been achieved by the player.